+++ runtime/moonbase/tests/common/mod.rs
@@ -1,0 +1,253 @@
+// Copyright 2019-2021 PureStake Inc.
+// This file is part of Moonbeam.
+
+// Moonbeam is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Moonbeam is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
+
+#![allow(dead_code)]
+
+use cumulus_primitives_parachain_inherent::ParachainInherentData;
+use frame_support::{
+	assert_ok,
+	dispatch::Dispatchable,
+	traits::{GenesisBuild, OnFinalize, OnInitialize},
+};
+pub use moonbase_runtime::{
+	currency::UNITS, AccountId, AuthorInherent, Balance, Balances, Call, CrowdloanRewards,
+	Ethereum, Event, Executive, FixedGasPrice, InflationInfo, ParachainStaking, Range, Runtime,
+	System, TransactionConverter, UncheckedExtrinsic,
+};
+use nimbus_primitives::NimbusId;
+use pallet_evm::GenesisAccount;
+use sp_core::{Public, H160};
+use sp_runtime::Perbill;
+
+use std::collections::BTreeMap;
+
+pub fn run_to_block(n: u32) {
+	while System::block_number() < n {
+		Ethereum::on_finalize(System::block_number());
+		AuthorInherent::on_finalize(System::block_number());
+		ParachainStaking::on_finalize(System::block_number());
+		System::set_block_number(System::block_number() + 1);
+		AuthorInherent::on_initialize(System::block_number());
+		Ethereum::on_initialize(System::block_number());
+	}
+}
+
+pub fn last_event() -> Event {
+	System::events().pop().expect("Event expected").event
+}
+
+pub struct ExtBuilder {
+	// endowed accounts with balances
+	balances: Vec<(AccountId, Balance)>,
+	// [collator, amount]
+	collators: Vec<(AccountId, Balance)>,
+	// [nominator, collator, nomination_amount]
+	nominators: Vec<(AccountId, AccountId, Balance)>,
+	// per-round inflation config
+	inflation: InflationInfo<Balance>,
+	// Crowdloan fund
+	crowdloan_fund: Balance,
+	// Chain id
+	chain_id: u64,
+	// EVM genesis accounts
+	evm_accounts: BTreeMap<H160, GenesisAccount>,
+}
+
+impl Default for ExtBuilder {
+	fn default() -> ExtBuilder {
+		ExtBuilder {
+			balances: vec![],
+			nominators: vec![],
+			collators: vec![],
+			inflation: InflationInfo {
+				expect: Range {
+					min: 100_000 * UNITS,
+					ideal: 200_000 * UNITS,
+					max: 500_000 * UNITS,
+				},
+				// not used
+				annual: Range {
+					min: Perbill::from_percent(50),
+					ideal: Perbill::from_percent(50),
+					max: Perbill::from_percent(50),
+				},
+				// unrealistically high parameterization, only for testing
+				round: Range {
+					min: Perbill::from_percent(5),
+					ideal: Perbill::from_percent(5),
+					max: Perbill::from_percent(5),
+				},
+			},
+			crowdloan_fund: 0,
+			chain_id: CHAIN_ID,
+			evm_accounts: BTreeMap::new(),
+		}
+	}
+}
+
+impl ExtBuilder {
+	pub fn with_evm_accounts(mut self, accounts: BTreeMap<H160, GenesisAccount>) -> Self {
+		self.evm_accounts = accounts;
+		self
+	}
+
+	pub fn with_balances(mut self, balances: Vec<(AccountId, Balance)>) -> Self {
+		self.balances = balances;
+		self
+	}
+
+	pub fn with_collators(mut self, collators: Vec<(AccountId, Balance)>) -> Self {
+		self.collators = collators;
+		self
+	}
+
+	pub fn with_nominators(mut self, nominators: Vec<(AccountId, AccountId, Balance)>) -> Self {
+		self.nominators = nominators;
+		self
+	}
+
+	pub fn with_crowdloan_fund(mut self, crowdloan_fund: Balance) -> Self {
+		self.crowdloan_fund = crowdloan_fund;
+		self
+	}
+
+	pub fn with_inflation(mut self, inflation: InflationInfo<Balance>) -> Self {
+		self.inflation = inflation;
+		self
+	}
+
+	pub fn build(self) -> sp_io::TestExternalities {
+		let mut t = frame_system::GenesisConfig::default()
+			.build_storage::<Runtime>()
+			.unwrap();
+
+		pallet_balances::GenesisConfig::<Runtime> {
+			balances: self.balances,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		let mut stakers: Vec<(AccountId, Option<AccountId>, Balance)> = Vec::new();
+		for collator in self.collators {
+			stakers.push((collator.0, None, collator.1));
+		}
+		for nominator in self.nominators {
+			stakers.push((nominator.0, Some(nominator.1), nominator.2));
+		}
+		parachain_staking::GenesisConfig::<Runtime> {
+			stakers,
+			inflation_config: self.inflation,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		pallet_crowdloan_rewards::GenesisConfig::<Runtime> {
+			funded_amount: self.crowdloan_fund,
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		// Here we map the author id ALICE_NIMBUS to the AccountId ALICE
+		// This is not (currently) configureable because it is enough for all of our tests
+		// It could bemade configureable.
+		pallet_author_mapping::GenesisConfig::<Runtime> {
+			author_ids: vec![(NimbusId::from_slice(&ALICE_NIMBUS), AccountId::from(ALICE))],
+		}
+		.assimilate_storage(&mut t)
+		.unwrap();
+
+		<pallet_ethereum_chain_id::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
+			&pallet_ethereum_chain_id::GenesisConfig {
+				chain_id: self.chain_id,
+			},
+			&mut t,
+		)
+		.unwrap();
+
+		<pallet_evm::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
+			&pallet_evm::GenesisConfig {
+				accounts: self.evm_accounts,
+			},
+			&mut t,
+		)
+		.unwrap();
+
+		pallet_ethereum::GenesisConfig::assimilate_storage::<Runtime>(
+			&pallet_ethereum::GenesisConfig {},
+			&mut t,
+		)
+		.unwrap();
+
+		let mut ext = sp_io::TestExternalities::new(t);
+		ext.execute_with(|| System::set_block_number(1));
+		ext
+	}
+}
+
+pub const CHAIN_ID: u64 = 1281;
+pub const ALICE: [u8; 20] = [4u8; 20];
+pub const ALICE_NIMBUS: [u8; 32] = [4u8; 32];
+pub const BOB: [u8; 20] = [5u8; 20];
+pub const CHARLIE: [u8; 20] = [6u8; 20];
+pub const DAVE: [u8; 20] = [7u8; 20];
+pub const EVM_CONTRACT: [u8; 20] = [8u8; 20];
+
+pub fn origin_of(account_id: AccountId) -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::signed(account_id)
+}
+
+pub fn inherent_origin() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::none()
+}
+
+pub fn root_origin() -> <Runtime as frame_system::Config>::Origin {
+	<Runtime as frame_system::Config>::Origin::root()
+}
+
+/// Mock the inherent that sets author in `author-inherent`
+pub fn set_author(a: NimbusId) {
+	assert_ok!(
+		Call::AuthorInherent(pallet_author_inherent::Call::<Runtime>::set_author(a))
+			.dispatch(inherent_origin())
+	);
+}
+
+/// Mock the inherent that sets validation data in ParachainSystem, which
+/// contains the `relay_chain_block_number`, which is used in `author-filter` as a
+/// source of randomness to filter valid authors at each block.
+pub fn set_parachain_inherent_data() {
+	use cumulus_primitives_core::PersistedValidationData;
+	use cumulus_test_relay_sproof_builder::RelayStateSproofBuilder;
+	let (relay_parent_storage_root, relay_chain_state) =
+		RelayStateSproofBuilder::default().into_state_root_and_proof();
+	let vfp = PersistedValidationData {
+		relay_parent_number: 1u32,
+		relay_parent_storage_root,
+		..Default::default()
+	};
+	let parachain_inherent_data = ParachainInherentData {
+		validation_data: vfp,
+		relay_chain_state: relay_chain_state,
+		downward_messages: Default::default(),
+		horizontal_messages: Default::default(),
+	};
+	assert_ok!(Call::ParachainSystem(
+		cumulus_pallet_parachain_system::Call::<Runtime>::set_validation_data(
+			parachain_inherent_data
+		)
+	)
+	.dispatch(inherent_origin()));
+}
