+++ runtime/moonbase/tests/integration_test.rs
@@ -16,272 +16,17 @@
 
 //! Moonbase Runtime Integration Tests
 
-#![cfg(test)]
+mod common;
+use common::*;
 
-use cumulus_primitives_parachain_inherent::ParachainInherentData;
 use evm::{Context, ExitSucceed};
-use frame_support::{
-	assert_noop, assert_ok,
-	dispatch::Dispatchable,
-	traits::{GenesisBuild, OnFinalize, OnInitialize},
-};
-use moonbase_runtime::{
-	currency::UNITS, AccountId, AuthorInherent, Balance, Balances, Call, CrowdloanRewards,
-	Ethereum, Event, Executive, FixedGasPrice, InflationInfo, ParachainStaking, Range, Runtime,
-	System, TransactionConverter, UncheckedExtrinsic,
-};
+use frame_support::{assert_noop, assert_ok, dispatch::Dispatchable};
 use nimbus_primitives::NimbusId;
-use pallet_evm::{
-	Account as EVMAccount, AddressMapping, FeeCalculator, GenesisAccount, PrecompileSet,
-};
+use pallet_evm::PrecompileSet;
 use parachain_staking::Bond;
 use precompiles::MoonbeamPrecompiles;
-use sp_core::{Public, H160, H256, U256};
-use sp_runtime::{DispatchError, Perbill};
-
-use fp_rpc::runtime_decl_for_EthereumRuntimeRPCApi::EthereumRuntimeRPCApi;
-use fp_rpc::ConvertTransaction;
-use moonbeam_rpc_primitives_debug::runtime_decl_for_DebugRuntimeApi::DebugRuntimeApi;
-use moonbeam_rpc_primitives_txpool::runtime_decl_for_TxPoolRuntimeApi::TxPoolRuntimeApi;
-use std::collections::BTreeMap;
-use std::str::FromStr;
-
-fn run_to_block(n: u32) {
-	while System::block_number() < n {
-		Ethereum::on_finalize(System::block_number());
-		AuthorInherent::on_finalize(System::block_number());
-		ParachainStaking::on_finalize(System::block_number());
-		System::set_block_number(System::block_number() + 1);
-		AuthorInherent::on_initialize(System::block_number());
-		Ethereum::on_initialize(System::block_number());
-	}
-}
-
-fn last_event() -> Event {
-	System::events().pop().expect("Event expected").event
-}
-
-struct ExtBuilder {
-	// endowed accounts with balances
-	balances: Vec<(AccountId, Balance)>,
-	// [collator, amount]
-	collators: Vec<(AccountId, Balance)>,
-	// [nominator, collator, nomination_amount]
-	nominators: Vec<(AccountId, AccountId, Balance)>,
-	// per-round inflation config
-	inflation: InflationInfo<Balance>,
-	// Crowdloan fund
-	crowdloan_fund: Balance,
-	// Chain id
-	chain_id: u64,
-	// EVM genesis accounts
-	evm_accounts: BTreeMap<H160, GenesisAccount>,
-}
-
-impl Default for ExtBuilder {
-	fn default() -> ExtBuilder {
-		ExtBuilder {
-			balances: vec![],
-			nominators: vec![],
-			collators: vec![],
-			inflation: InflationInfo {
-				expect: Range {
-					min: 100_000 * UNITS,
-					ideal: 200_000 * UNITS,
-					max: 500_000 * UNITS,
-				},
-				// not used
-				annual: Range {
-					min: Perbill::from_percent(50),
-					ideal: Perbill::from_percent(50),
-					max: Perbill::from_percent(50),
-				},
-				// unrealistically high parameterization, only for testing
-				round: Range {
-					min: Perbill::from_percent(5),
-					ideal: Perbill::from_percent(5),
-					max: Perbill::from_percent(5),
-				},
-			},
-			crowdloan_fund: 0,
-			chain_id: CHAIN_ID,
-			evm_accounts: BTreeMap::new(),
-		}
-	}
-}
-
-impl ExtBuilder {
-	fn with_evm_accounts(mut self, accounts: BTreeMap<H160, GenesisAccount>) -> Self {
-		self.evm_accounts = accounts;
-		self
-	}
-
-	fn with_balances(mut self, balances: Vec<(AccountId, Balance)>) -> Self {
-		self.balances = balances;
-		self
-	}
-
-	fn with_collators(mut self, collators: Vec<(AccountId, Balance)>) -> Self {
-		self.collators = collators;
-		self
-	}
-
-	fn with_nominators(mut self, nominators: Vec<(AccountId, AccountId, Balance)>) -> Self {
-		self.nominators = nominators;
-		self
-	}
-
-	fn with_crowdloan_fund(mut self, crowdloan_fund: Balance) -> Self {
-		self.crowdloan_fund = crowdloan_fund;
-		self
-	}
-
-	#[allow(dead_code)]
-	fn with_inflation(mut self, inflation: InflationInfo<Balance>) -> Self {
-		self.inflation = inflation;
-		self
-	}
-
-	fn build(self) -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default()
-			.build_storage::<Runtime>()
-			.unwrap();
-
-		pallet_balances::GenesisConfig::<Runtime> {
-			balances: self.balances,
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		let mut stakers: Vec<(AccountId, Option<AccountId>, Balance)> = Vec::new();
-		for collator in self.collators {
-			stakers.push((collator.0, None, collator.1));
-		}
-		for nominator in self.nominators {
-			stakers.push((nominator.0, Some(nominator.1), nominator.2));
-		}
-		parachain_staking::GenesisConfig::<Runtime> {
-			stakers,
-			inflation_config: self.inflation,
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		pallet_crowdloan_rewards::GenesisConfig::<Runtime> {
-			funded_amount: self.crowdloan_fund,
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		// Here we map the author id ALICE_NIMBUS to the AccountId ALICE
-		// This is not (currently) configureable because it is enough for all of our tests
-		// It could bemade configureable.
-		pallet_author_mapping::GenesisConfig::<Runtime> {
-			author_ids: vec![(NimbusId::from_slice(&ALICE_NIMBUS), AccountId::from(ALICE))],
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		<pallet_ethereum_chain_id::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
-			&pallet_ethereum_chain_id::GenesisConfig {
-				chain_id: self.chain_id,
-			},
-			&mut t,
-		)
-		.unwrap();
-
-		<pallet_evm::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
-			&pallet_evm::GenesisConfig {
-				accounts: self.evm_accounts,
-			},
-			&mut t,
-		)
-		.unwrap();
-
-		pallet_ethereum::GenesisConfig::assimilate_storage::<Runtime>(
-			&pallet_ethereum::GenesisConfig {},
-			&mut t,
-		)
-		.unwrap();
-
-		let mut ext = sp_io::TestExternalities::new(t);
-		ext.execute_with(|| System::set_block_number(1));
-		ext
-	}
-}
-
-const CHAIN_ID: u64 = 1281;
-const ALICE: [u8; 20] = [4u8; 20];
-const ALICE_NIMBUS: [u8; 32] = [4u8; 32];
-const BOB: [u8; 20] = [5u8; 20];
-const CHARLIE: [u8; 20] = [6u8; 20];
-const DAVE: [u8; 20] = [7u8; 20];
-const EVM_CONTRACT: [u8; 20] = [8u8; 20];
-
-fn origin_of(account_id: AccountId) -> <Runtime as frame_system::Config>::Origin {
-	<Runtime as frame_system::Config>::Origin::signed(account_id)
-}
-
-fn inherent_origin() -> <Runtime as frame_system::Config>::Origin {
-	<Runtime as frame_system::Config>::Origin::none()
-}
-
-fn root_origin() -> <Runtime as frame_system::Config>::Origin {
-	<Runtime as frame_system::Config>::Origin::root()
-}
-
-/// Mock the inherent that sets author in `author-inherent`
-fn set_author(a: NimbusId) {
-	assert_ok!(
-		Call::AuthorInherent(pallet_author_inherent::Call::<Runtime>::set_author(a))
-			.dispatch(inherent_origin())
-	);
-}
-
-/// Mock the inherent that sets validation data in ParachainSystem, which
-/// contains the `relay_chain_block_number`, which is used in `author-filter` as a
-/// source of randomness to filter valid authors at each block.
-fn set_parachain_inherent_data() {
-	use cumulus_primitives_core::PersistedValidationData;
-	use cumulus_test_relay_sproof_builder::RelayStateSproofBuilder;
-	let (relay_parent_storage_root, relay_chain_state) =
-		RelayStateSproofBuilder::default().into_state_root_and_proof();
-	let vfp = PersistedValidationData {
-		relay_parent_number: 1u32,
-		relay_parent_storage_root,
-		..Default::default()
-	};
-	let parachain_inherent_data = ParachainInherentData {
-		validation_data: vfp,
-		relay_chain_state: relay_chain_state,
-		downward_messages: Default::default(),
-		horizontal_messages: Default::default(),
-	};
-	assert_ok!(Call::ParachainSystem(
-		cumulus_pallet_parachain_system::Call::<Runtime>::set_validation_data(
-			parachain_inherent_data
-		)
-	)
-	.dispatch(inherent_origin()));
-}
-
-fn ethereum_transaction() -> pallet_ethereum::Transaction {
-	// {from: 0x6be02d1d3665660d22ff9624b7be0551ee1ac91b, .., gasPrice: "0x01"}
-	let bytes = hex::decode(
-		"f86880843b9aca0083b71b0094111111111111111111111111111111111111111182020080820a26a0\
-		8c69faf613b9f72dbb029bb5d5acf42742d214c79743507e75fdc8adecdee928a001be4f58ff278ac61\
-		125a81a582a717d9c5d6554326c01b878297c6522b12282",
-	)
-	.expect("Transaction bytes.");
-	let transaction = rlp::decode::<pallet_ethereum::Transaction>(&bytes[..]);
-	assert!(transaction.is_ok());
-	transaction.unwrap()
-}
-
-fn uxt() -> UncheckedExtrinsic {
-	let converter = TransactionConverter;
-	converter.convert_transaction(ethereum_transaction())
-}
+use sp_core::{Public, H160, U256};
+use sp_runtime::DispatchError;
 
 #[test]
 fn join_collator_candidates() {
@@ -1360,320 +1105,4 @@ fn min_nomination_via_precompile() {
 			expected_result
 		);
 	});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_chain_id() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_eq!(Runtime::chain_id(), CHAIN_ID);
-	});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_account_basic() {
-	ExtBuilder::default()
-		.with_balances(vec![(AccountId::from(ALICE), 2_000 * UNITS)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(
-				Runtime::account_basic(H160::from(ALICE)),
-				EVMAccount {
-					balance: U256::from(2_000 * UNITS),
-					nonce: U256::zero()
-				}
-			);
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_gas_price() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_eq!(Runtime::gas_price(), FixedGasPrice::min_gas_price());
-	});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_account_code_at() {
-	let address = H160::from(EVM_CONTRACT);
-	let code: Vec<u8> = vec![1, 2, 3, 4, 5];
-	ExtBuilder::default()
-		.with_evm_accounts({
-			let mut map = BTreeMap::new();
-			map.insert(
-				address,
-				GenesisAccount {
-					balance: U256::zero(),
-					code: code.clone(),
-					nonce: Default::default(),
-					storage: Default::default(),
-				},
-			);
-			map
-		})
-		.build()
-		.execute_with(|| {
-			assert_eq!(Runtime::account_code_at(address), code);
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_author() {
-	ExtBuilder::default()
-		.with_collators(vec![(AccountId::from(ALICE), 1_000 * UNITS)])
-		.with_balances(vec![
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.with_nominators(vec![(
-			AccountId::from(BOB),
-			AccountId::from(ALICE),
-			500 * UNITS,
-		)])
-		.build()
-		.execute_with(|| {
-			set_parachain_inherent_data();
-			set_author(NimbusId::from_slice(&ALICE_NIMBUS));
-			assert_eq!(Runtime::author(), H160::from(ALICE));
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_storage_at() {
-	let address = H160::from(EVM_CONTRACT);
-	let mut key = [0u8; 32];
-	key[31..32].copy_from_slice(&[6u8][..]);
-	let mut value = [0u8; 32];
-	value[31..32].copy_from_slice(&[7u8][..]);
-	let item = H256::from_slice(&key[..]);
-	let mut storage: BTreeMap<H256, H256> = BTreeMap::new();
-	storage.insert(H256::from_slice(&key[..]), item);
-	ExtBuilder::default()
-		.with_evm_accounts({
-			let mut map = BTreeMap::new();
-			map.insert(
-				address,
-				GenesisAccount {
-					balance: U256::zero(),
-					code: Vec::new(),
-					nonce: Default::default(),
-					storage: storage.clone(),
-				},
-			);
-			map
-		})
-		.build()
-		.execute_with(|| {
-			assert_eq!(Runtime::storage_at(address, U256::from(6)), item);
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_call() {
-	ExtBuilder::default()
-		.with_balances(vec![
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 2_000 * UNITS),
-		])
-		.build()
-		.execute_with(|| {
-			let execution_result = Runtime::call(
-				H160::from(ALICE),  // from
-				H160::from(BOB),    // to
-				Vec::new(),         // data
-				U256::from(1000),   // value
-				U256::from(100000), // gas_limit
-				None,               // gas_price
-				None,               // nonce
-				false,              // estimate
-			);
-			assert!(execution_result.is_ok());
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_create() {
-	ExtBuilder::default()
-		.with_balances(vec![(AccountId::from(ALICE), 2_000 * UNITS)])
-		.build()
-		.execute_with(|| {
-			let execution_result = Runtime::create(
-				H160::from(ALICE),  // from
-				vec![0, 1, 1, 0],   // data
-				U256::zero(),       // value
-				U256::from(100000), // gas_limit
-				None,               // gas_price
-				None,               // nonce
-				false,              // estimate
-			);
-			assert!(execution_result.is_ok());
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_current_transaction_statuses() {
-	let alith = <Runtime as pallet_evm::Config>::AddressMapping::into_account_id(
-		H160::from_str("6be02d1d3665660d22ff9624b7be0551ee1ac91b")
-			.expect("internal H160 is valid; qed"),
-	);
-	ExtBuilder::default()
-		.with_collators(vec![(AccountId::from(ALICE), 1_000 * UNITS)])
-		.with_balances(vec![
-			(alith, 2_000 * UNITS),
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.with_nominators(vec![(
-			AccountId::from(BOB),
-			AccountId::from(ALICE),
-			500 * UNITS,
-		)])
-		.build()
-		.execute_with(|| {
-			set_parachain_inherent_data();
-			set_author(NimbusId::from_slice(&ALICE_NIMBUS));
-			// {from: 0x6be02d1d3665660d22ff9624b7be0551ee1ac91b, .., gasPrice: "0x01"}
-			let bytes = hex::decode(
-				"f86880843b9aca0083b71b0094111111111111111111111111111111111111111182020080820a26a0\
-				8c69faf613b9f72dbb029bb5d5acf42742d214c79743507e75fdc8adecdee928a001be4f58ff278ac61\
-				125a81a582a717d9c5d6554326c01b878297c6522b12282")
-				.expect("Transaction bytes.");
-			let transaction = rlp::decode::<pallet_ethereum::Transaction>(&bytes[..]);
-			assert!(transaction.is_ok());
-			let converter = TransactionConverter;
-			let uxt: UncheckedExtrinsic =
-				converter.convert_transaction(transaction.unwrap().clone());
-			let _result = Executive::apply_extrinsic(uxt);
-			run_to_block(2);
-			let statuses =
-				Runtime::current_transaction_statuses().expect("Transaction statuses result.");
-			assert_eq!(statuses.len(), 1);
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_current_block() {
-	ExtBuilder::default()
-		.with_collators(vec![(AccountId::from(ALICE), 1_000 * UNITS)])
-		.with_balances(vec![
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.with_nominators(vec![(
-			AccountId::from(BOB),
-			AccountId::from(ALICE),
-			500 * UNITS,
-		)])
-		.build()
-		.execute_with(|| {
-			set_parachain_inherent_data();
-			set_author(NimbusId::from_slice(&ALICE_NIMBUS));
-			run_to_block(2);
-			let block = Runtime::current_block().expect("Block result.");
-			assert_eq!(block.header.number, U256::from(1));
-		});
-}
-
-#[test]
-fn ethereum_runtime_rpc_api_current_receipts() {
-	let alith = <Runtime as pallet_evm::Config>::AddressMapping::into_account_id(
-		H160::from_str("6be02d1d3665660d22ff9624b7be0551ee1ac91b")
-			.expect("internal H160 is valid; qed"),
-	);
-	ExtBuilder::default()
-		.with_collators(vec![(AccountId::from(ALICE), 1_000 * UNITS)])
-		.with_balances(vec![
-			(alith, 2_000 * UNITS),
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.with_nominators(vec![(
-			AccountId::from(BOB),
-			AccountId::from(ALICE),
-			500 * UNITS,
-		)])
-		.build()
-		.execute_with(|| {
-			set_parachain_inherent_data();
-			set_author(NimbusId::from_slice(&ALICE_NIMBUS));
-			let _result = Executive::apply_extrinsic(uxt());
-			run_to_block(2);
-			let receipts = Runtime::current_receipts().expect("Receipts result.");
-			assert_eq!(receipts.len(), 1);
-		});
-}
-
-#[test]
-fn txpool_runtime_api_extrinsic_filter() {
-	ExtBuilder::default().build().execute_with(|| {
-		let non_eth_uxt = UncheckedExtrinsic::new_unsigned(
-			pallet_balances::Call::<Runtime>::transfer(AccountId::from(BOB), 1 * UNITS).into(),
-		);
-		let eth_uxt = uxt();
-		let txpool = Runtime::extrinsic_filter(
-			vec![eth_uxt.clone(), non_eth_uxt.clone()],
-			vec![uxt(), non_eth_uxt],
-		);
-		assert_eq!(txpool.ready.len(), 1);
-		assert_eq!(txpool.future.len(), 1);
-	});
-}
-
-#[test]
-fn debug_runtime_api_trace_transaction() {
-	let alith = <Runtime as pallet_evm::Config>::AddressMapping::into_account_id(
-		H160::from_str("6be02d1d3665660d22ff9624b7be0551ee1ac91b")
-			.expect("internal H160 is valid; qed"),
-	);
-	ExtBuilder::default()
-		.with_balances(vec![
-			(alith, 2_000 * UNITS),
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.build()
-		.execute_with(|| {
-			let non_eth_uxt = UncheckedExtrinsic::new_unsigned(
-				pallet_balances::Call::<Runtime>::transfer(AccountId::from(BOB), 1 * UNITS).into(),
-			);
-			let transaction = ethereum_transaction();
-			let eth_uxt = uxt();
-			assert!(Runtime::trace_transaction(
-				vec![non_eth_uxt.clone(), eth_uxt, non_eth_uxt.clone()],
-				&transaction,
-				moonbeam_rpc_primitives_debug::single::TraceType::Raw {
-					disable_storage: true,
-					disable_memory: true,
-					disable_stack: true,
-				}
-			)
-			.is_ok());
-		});
-}
-
-#[test]
-fn debug_runtime_api_trace_block() {
-	let alith = <Runtime as pallet_evm::Config>::AddressMapping::into_account_id(
-		H160::from_str("6be02d1d3665660d22ff9624b7be0551ee1ac91b")
-			.expect("internal H160 is valid; qed"),
-	);
-	ExtBuilder::default()
-		.with_balances(vec![
-			(alith, 2_000 * UNITS),
-			(AccountId::from(ALICE), 2_000 * UNITS),
-			(AccountId::from(BOB), 1_000 * UNITS),
-		])
-		.build()
-		.execute_with(|| {
-			let non_eth_uxt = UncheckedExtrinsic::new_unsigned(
-				pallet_balances::Call::<Runtime>::transfer(AccountId::from(BOB), 1 * UNITS).into(),
-			);
-			let eth_uxt = uxt();
-			assert!(Runtime::trace_block(vec![
-				non_eth_uxt.clone(),
-				eth_uxt.clone(),
-				non_eth_uxt,
-				eth_uxt
-			],)
-			.is_ok());
-		});
 }
